# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Naturascan API",
        "description": "Naturascan API",
        "contact": {
            "name": "Gilles",
            "email": "gillesakakpo01@gmail.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/login": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Se connecter",
                "operationId": "authLogin",
                "requestBody": {
                    "description": "Champs d'authentification",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "utilisateur@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "motdepasse"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Connexion réussie",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "access_token": {
                                            "type": "string"
                                        },
                                        "token_type": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "202": {
                        "description": "Compte bloqué",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Erreurs de validation"
                    }
                }
            }
        },
        "/api/logout": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Se déconnecter",
                "operationId": "authLogout",
                "responses": {
                    "200": {
                        "description": "Déconnexion réussie"
                    },
                    "401": {
                        "description": "Non autorisé"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/me": {
            "get": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Obtenir les informations de l'utilisateur authentifié",
                "operationId": "authMe",
                "responses": {
                    "200": {
                        "description": "Récupération réussie des informations de l'utilisateur authentifié",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "access_token": {
                                            "type": "string"
                                        },
                                        "token_type": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non autorisé",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Token invalide"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/register": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Inscription d'un nouvel utilisateur",
                "operationId": "authRegister",
                "requestBody": {
                    "description": "Champs d'inscription d'un nouvel utilisateur",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "John"
                                    },
                                    "firstname": {
                                        "type": "string",
                                        "example": "Doe"
                                    },
                                    "email": {
                                        "description": "Adresse email unique de l'utilisateur",
                                        "type": "string",
                                        "format": "email",
                                        "example": "john@yopmail.com"
                                    },
                                    "password": {
                                        "description": "Mot de passe de l'utilisateur (min. 8 caractères)",
                                        "type": "string",
                                        "format": "password",
                                        "example": "12345678"
                                    },
                                    "password_confirmation": {
                                        "description": "Mot de passe de l'utilisateur (min. 8 caractères)",
                                        "type": "string",
                                        "format": "password",
                                        "example": "12345678"
                                    },
                                    "mobile_number": {
                                        "description": "Téléphone",
                                        "type": "string",
                                        "example": "02056852"
                                    },
                                    "adress": {
                                        "description": "Adresse",
                                        "type": "string",
                                        "example": "johndoe"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Inscription réussie",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "access_token": {
                                            "type": "string",
                                            "example": "votre_access_token"
                                        },
                                        "token_type": {
                                            "type": "string",
                                            "example": "Bearer"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Erreurs de validation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Erreur de validations"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/observers": {
            "get": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Obtenir la liste des observateurs",
                "operationId": "getObservers",
                "responses": {
                    "200": {
                        "description": "Liste des observateurs",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/User"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "User": {
                "title": "User",
                "description": "User",
                "properties": {
                    "id": {
                        "title": "id",
                        "type": "integer"
                    },
                    "name": {
                        "title": "name",
                        "type": "string"
                    },
                    "firstname": {
                        "title": "firstname",
                        "type": "string"
                    },
                    "email": {
                        "title": "email",
                        "type": "string"
                    },
                    "mobile_number": {
                        "title": "mobile_number",
                        "type": "string"
                    },
                    "adress": {
                        "title": "adress",
                        "type": "string"
                    },
                    "access": {
                        "title": "access",
                        "type": "string"
                    },
                    "pseudo": {
                        "title": "pseudo",
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}